+++ programs/amm/src/endpoints/farming.rs
@@ -1,7 +1,11 @@
 pub mod add_harvest;
 pub mod create_farm;
 pub mod remove_harvest;
+pub mod set_min_snapshot_window;
+pub mod take_snapshot;
 
 pub use add_harvest::*;
 pub use create_farm::*;
 pub use remove_harvest::*;
+pub use set_min_snapshot_window::*;
+pub use take_snapshot::*;